+++ lib/big5_2003.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2005 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2005, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -180,7 +180,7 @@ static const unsigned short big5_2003_2uni_pagef9[41] = {
 };
 
 static int
-big5_2003_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+big5_2003_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c = *s;
   /* Code set 0 (ASCII) */
@@ -267,7 +267,7 @@ static const unsigned char big5_2003_2charset_page25[29] = {
 };
 
 static int
-big5_2003_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+big5_2003_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   unsigned char buf[2];
   int ret;
