+++ lib/cns11643_inv.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2005, 2012 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2005, 2012, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -15365,7 +15365,7 @@ static const Summary16 cns11643_inv_uni2indx_page2f8[34] = {
 };
 
 static int
-cns11643_inv_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cns11643_inv_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   if (n >= 2) {
     const Summary16 *summary = NULL;
