+++ lib/cp50221_0208_ext.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2012 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2012, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -43,7 +43,7 @@ static const unsigned short cp50221_0208_ext_2uni[96] = {
 };
 
 static int
-cp50221_0208_ext_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+cp50221_0208_ext_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c = *s;
   if (c < 0x60) {
@@ -122,7 +122,7 @@ static const unsigned char cp50221_0208_ext_page33[208] = {
 };
 
 static int
-cp50221_0208_ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cp50221_0208_ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   unsigned char c = 0;
   if (wc >= 0x2110 && wc < 0x2170)
