+++ lib/cp863.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2002 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2002, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -49,7 +49,7 @@ static const unsigned short cp863_2uni[128] = {
 };
 
 static int
-cp863_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+cp863_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c = *s;
   if (c < 0x80)
@@ -124,7 +124,7 @@ static const unsigned char cp863_page25[168] = {
 };
 
 static int
-cp863_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cp863_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   unsigned char c = 0;
   if (wc < 0x0080) {
