+++ lib/cp932ext.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2012 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2012, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -145,7 +145,7 @@ static const unsigned short cp932ext_2uni_pagefa[388] = {
 };
 
 static int
-cp932ext_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+cp932ext_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c1 = s[0];
   if ((c1 == 0x87) || (c1 >= 0xed && c1 <= 0xee) || (c1 >= 0xfa && c1 <= 0xfc)) {
@@ -653,7 +653,7 @@ static const Summary16 cp932ext_uni2indx_pageff[15] = {
 };
 
 static int
-cp932ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cp932ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   if (n >= 2) {
     const Summary16 *summary = NULL;
