+++ lib/cp936.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2005 Free Software Foundation, Inc.
+ * Copyright (C) 2005, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -47,7 +47,7 @@
  */
 
 static int
-cp936_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+cp936_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   /* Try GBK first. */
   {
@@ -89,7 +89,7 @@ cp936_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
 }
 
 static int
-cp936_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cp936_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   /* Try GBK first. */
   {
