+++ lib/cp950ext.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2012 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2012, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -32,7 +32,7 @@ static const unsigned short cp950ext_2uni_pagef9[157-116] = {
 };
 
 static int
-cp950ext_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+cp950ext_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c1 = s[0];
   if ((c1 == 0xf9)) {
@@ -117,7 +117,7 @@ static const Summary16 cp950ext_uni2indx_page92[12] = {
 };
 
 static int
-cp950ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+cp950ext_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   if (n >= 2) {
     const Summary16 *summary = NULL;
