+++ lib/iso2022_jpms.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2008, 2011-2012 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2008, 2011-2012, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -81,7 +81,7 @@
 #define STATE_JISX0212MS        4  /* Esc $ ( D */
 
 static int
-iso2022_jpms_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+iso2022_jpms_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   state_t state = conv->istate;
   int count = 0;
@@ -278,7 +278,7 @@ ilseq:
 }
 
 static int
-iso2022_jpms_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+iso2022_jpms_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   state_t state = conv->ostate;
   unsigned char buf[2];
@@ -423,7 +423,7 @@ iso2022_jpms_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 }
 
 static int
-iso2022_jpms_reset (conv_t conv, unsigned char *r, int n)
+iso2022_jpms_reset (conv_t conv, unsigned char *r, size_t n)
 {
   state_t state = conv->ostate;
   if (state != STATE_ASCII) {
